character_interaction/requirements2.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
assemblyai==0.30.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
cmudict==1.0.27
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
distro==1.9.0
Django==5.0.1
django-allauth==0.60.1
django-appconf==1.0.6
django-image-cropping==1.7
djangorestframework==3.15.1
easy-thumbnails==2.8.5
elevenlabs==0.3.0b0
executing==2.0.1
fake-bpy-module==20240712
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
fpdf==1.7.2
fsspec==2024.6.1
g2p-en==2.1.0
gast==0.5.4
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.16.5
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.3
gTTS==2.5.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.2.0
importlib_resources==6.4.0
incremental==24.7.2
inflect==7.3.1
ipython==8.26.0
jedi==0.19.1
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.4.0
namex==0.0.8
nltk==3.8.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.14
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.1
parso==0.8.4
pillow==10.2.0
prompt-toolkit==3.0.43
pronouncing==0.2.0
proto-plus==1.24.0
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0

